+++ packages/blade/src/components/ActionList/styles/StyledActionListItem.native.tsx
@@ -6,6 +6,7 @@ import { getBaseActionListItemStyles } from './getBaseActionListItemStyles';
 const StyledActionListItem = styled(TouchableOpacity)<StyledActionListItemProps>((props) => {
   return {
     ...getBaseActionListItemStyles(props),
+    display: props.isVisible ? 'flex' : 'none',
     // React Native specific styles
     backgroundColor:
       props.isSelected && props.selectionType === 'single'