+++ packages/blade/src/components/Dropdown/Dropdown.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { DropdownContext } from './useDropdown';
 import type { DropdownContextType } from './useDropdown';
-import { componentIds } from './dropdownUtils';
 import type { DropdownProps } from './types';
+import { dropdownComponentIds } from './dropdownComponentIds';
 import { useId } from '~utils/useId';
 import { ComponentIds as bottomSheetComponentIds } from '~components/BottomSheet/componentIds';
 import { BottomSheetAndDropdownGlueContext } from '~components/BottomSheet/BottomSheetContext';
@@ -10,14 +10,17 @@ import { getStyledProps } from '~components/Box/styledProps';
 import BaseBox from '~components/Box/BaseBox';
 import { assignWithoutSideEffects } from '~utils/assignWithoutSideEffects';
 import { getComponentId, isValidAllowedChildren } from '~utils/isValidAllowedChildren';
+import { isReactNative } from '~utils';
 import { MetaConstants, metaAttribute } from '~utils/metaAttribute';
 import { throwBladeError } from '~utils/logger';
+import type { ContainerElementType } from '~utils/types';
 
 const validDropdownChildren = [
-  componentIds.triggers.SelectInput,
-  componentIds.triggers.DropdownButton,
-  componentIds.triggers.DropdownLink,
-  componentIds.DropdownOverlay,
+  dropdownComponentIds.triggers.SelectInput,
+  dropdownComponentIds.triggers.DropdownButton,
+  dropdownComponentIds.triggers.DropdownLink,
+  dropdownComponentIds.DropdownOverlay,
+  dropdownComponentIds.triggers.AutoComplete,
   bottomSheetComponentIds.BottomSheet,
 ];
 
@@ -56,6 +59,7 @@ const _Dropdown = ({
 }: DropdownProps): React.ReactElement => {
   const [isOpen, setIsOpen] = React.useState(false);
   const [options, setOptions] = React.useState<DropdownContextType['options']>([]);
+  const [filteredValues, setFilteredValues] = React.useState<string[]>([]);
   const [selectedIndices, setSelectedIndices] = React.useState<
     DropdownContextType['selectedIndices']
   >([]);
@@ -63,17 +67,13 @@ const _Dropdown = ({
     DropdownContextType['selectedIndices']
   >([]);
   const [activeIndex, setActiveIndex] = React.useState(-1);
-  const [shouldIgnoreBlur, setShouldIgnoreBlur] = React.useState(false);
+  const [activeTagIndex, setActiveTagIndex] = React.useState(-1);
   const [shouldIgnoreBlurAnimation, setShouldIgnoreBlurAnimation] = React.useState(false);
-  const triggererRef = React.useRef<HTMLButtonElement>(null);
-  /**
-   * In inputs, actual input is smaller than the visible input wrapper.
-   * You can set this reference in such cases so floating ui calculations happen correctly
-   * */
-  const triggererWrapperRef = React.useRef<HTMLDivElement>(null);
-  const actionListItemRef = React.useRef<HTMLDivElement>(null);
   const [hasFooterAction, setHasFooterAction] = React.useState(false);
-  const [hasLabelOnLeft, setHasLabelOnLeft] = React.useState(false);
+  const [
+    hasAutoCompleteInBottomSheetHeader,
+    setHasAutoCompleteInBottomSheetHeader,
+  ] = React.useState(false);
   const [isKeydownPressed, setIsKeydownPressed] = React.useState(false);
   const [changeCallbackTriggerer, setChangeCallbackTriggerer] = React.useState<
     DropdownContextType['changeCallbackTriggerer']
@@ -82,10 +82,20 @@ const _Dropdown = ({
   // keep track if dropdown contains bottomsheet
   const [dropdownHasBottomSheet, setDropdownHasBottomSheet] = React.useState(false);
 
-  const dropdownBaseId = useId('dropdown');
-
+  /**
+   * In inputs, actual input is smaller than the visible input wrapper.
+   * You can set this reference in such cases so floating ui calculations happen correctly
+   * */
+  const triggererWrapperRef = React.useRef<ContainerElementType>(null);
+  const triggererRef = React.useRef<HTMLButtonElement>(null);
+  const actionListItemRef = React.useRef<HTMLDivElement>(null);
   const dropdownTriggerer = React.useRef<DropdownContextType['dropdownTriggerer']>();
   const isFirstRenderRef = React.useRef(true);
+  const isTagDismissedRef = React.useRef<{ value: boolean } | null>({ value: false });
+  const visibleTagsCountRef = React.useRef<{ value: number }>({ value: 0 });
+  const dropdownContainerRef = React.useRef<HTMLDivElement>(null);
+
+  const dropdownBaseId = useId('dropdown');
 
   React.useEffect(() => {
     // Ignoring the `onDismiss` call on first render
@@ -101,6 +111,7 @@ const _Dropdown = ({
   }, [isOpen]);
 
   const close = React.useCallback(() => {
+    setActiveTagIndex(-1);
     setIsOpen(false);
     onDismiss?.();
   }, [onDismiss]);
@@ -118,12 +129,16 @@ const _Dropdown = ({
         }
       }
 
-      if (isValidAllowedChildren(child, componentIds.triggers.SelectInput)) {
+      if (isValidAllowedChildren(child, dropdownComponentIds.triggers.SelectInput)) {
         dropdownTriggerer.current = 'SelectInput';
       }
 
-      if (isValidAllowedChildren(child, componentIds.triggers.DropdownButton)) {
+      if (isValidAllowedChildren(child, dropdownComponentIds.triggers.DropdownButton)) {
         dropdownTriggerer.current = 'DropdownButton';
+      }
+
+      if (isValidAllowedChildren(child, dropdownComponentIds.triggers.AutoComplete)) {
+        dropdownTriggerer.current = 'AutoComplete';
       }
     }
   });
@@ -139,10 +154,13 @@ const _Dropdown = ({
       setControlledValueIndices,
       options,
       setOptions,
+      filteredValues,
+      setFilteredValues,
       activeIndex,
       setActiveIndex,
-      shouldIgnoreBlur,
-      setShouldIgnoreBlur,
+      activeTagIndex,
+      setActiveTagIndex,
+      visibleTagsCountRef,
       shouldIgnoreBlurAnimation,
       setShouldIgnoreBlurAnimation,
       isKeydownPressed,
@@ -154,13 +172,14 @@ const _Dropdown = ({
       selectionType,
       hasFooterAction,
       setHasFooterAction,
-      hasLabelOnLeft,
-      setHasLabelOnLeft,
+      hasAutoCompleteInBottomSheetHeader,
+      setHasAutoCompleteInBottomSheetHeader,
       dropdownTriggerer: dropdownTriggerer.current,
       changeCallbackTriggerer,
       setChangeCallbackTriggerer,
       isControlled,
       setIsControlled,
+      isTagDismissedRef,
     }),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [
@@ -168,12 +187,12 @@ const _Dropdown = ({
       selectedIndices,
       controlledValueIndices,
       options,
+      filteredValues,
       activeIndex,
-      shouldIgnoreBlur,
+      activeTagIndex,
       shouldIgnoreBlurAnimation,
       selectionType,
       hasFooterAction,
-      hasLabelOnLeft,
       isKeydownPressed,
       changeCallbackTriggerer,
       isControlled,
@@ -184,17 +203,66 @@ const _Dropdown = ({
     return {
       isOpen,
       dropdownHasBottomSheet,
+      hasAutoCompleteInBottomSheetHeader,
       setDropdownHasBottomSheet,
       // This is the dismiss function which will be injected into the BottomSheet
       // Basically <BottomSheet onDismiss={onBottomSheetDismiss} />
       onBottomSheetDismiss: close,
     };
-  }, [dropdownHasBottomSheet, isOpen, close]);
+  }, [dropdownHasBottomSheet, hasAutoCompleteInBottomSheetHeader, isOpen, close]);
+
+  React.useEffect((): (() => void) | undefined => {
+    if (!isReactNative()) {
+      const dropdown = dropdownContainerRef.current;
+
+      const documentClickHandler = (e: MouseEvent): void => {
+        const target = e.target as HTMLDivElement;
+
+        if (!target || !dropdown) {
+          return;
+        }
+
+        if (!dropdown.contains(target) && !isTagDismissedRef.current?.value) {
+          close();
+        }
+
+        if (isTagDismissedRef.current?.value) {
+          isTagDismissedRef.current.value = false;
+        }
+      };
+
+      const documentFocusHandler = (e: FocusEvent): void => {
+        const target = e.relatedTarget as HTMLDivElement;
+        setActiveIndex(-1);
+
+        if (!dropdown || !target) {
+          return;
+        }
+
+        if (!dropdown.contains(target)) {
+          close();
+        }
+      };
+
+      document.addEventListener('click', documentClickHandler);
+      document.addEventListener('focusout', documentFocusHandler);
+
+      return (): void => {
+        document.removeEventListener('click', documentClickHandler);
+        document.removeEventListener('focusout', documentFocusHandler);
+      };
+    }
+
+    return undefined;
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <BottomSheetAndDropdownGlueContext.Provider value={BottomSheetAndDropdownGlueContextValue}>
       <DropdownContext.Provider value={contextValue}>
         <BaseBox
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          ref={dropdownContainerRef as any}
           {...metaAttribute({ name: MetaConstants.Dropdown, testID })}
           {...getStyledProps(styledProps)}
         >
@@ -207,6 +275,8 @@ const _Dropdown = ({
   );
 };
 
-const Dropdown = assignWithoutSideEffects(_Dropdown, { componentId: componentIds.Dropdown });
+const Dropdown = assignWithoutSideEffects(_Dropdown, {
+  componentId: dropdownComponentIds.Dropdown,
+});
 
 export { Dropdown };