+++ packages/blade/src/components/Input/BaseInput/StyledBaseInput.web.tsx
@@ -25,6 +25,8 @@ const getWebInputStyles = (
       trailingIcon: props.trailingIcon,
       textAlign: props.textAlign,
       isTextArea: props.isTextArea,
+      hasTags: props.hasTags,
+      isDropdownTrigger: props.isDropdownTrigger,
     }),
     outline: 'none',
     border: 'none',